requirements.txt
asgiref==3.4.1

cx-Freeze==6.6
cx-Logging==3.0

decorator==4.4.1
defusedxml==0.6.0
Django==3.2.6
dj-static==0.0.6
django-tinymce==3.4.0
gunicorn==20.1.0

oauth2client==4.1.3
oauthlib==3.1.0
Pillow==7.0.0
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2

Send2Trash==1.5.0
six==1.13.0
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
django-autoslug==1.9.8


urllib3==1.26.4
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xlrd==2.0.1
zipp==3.4.1
